Create main
Ved Yog Charitable Trust
The main purpose of the trust is — To study, teach, and propagate Vedas, Vedic Scriptures, and Patanjali Yoga.

So that human-beings can achieve physical, spiritual, and social excellence, in their lives. Let the human-beings in the world shall be able to achieve happiness in this world as well as in the next world, i.e. in their future births. The main cause of happiness lies in the true knowledge of the GOD, the Soul, and Matter. This knowledge was given by the Supreme Lord at the beginning of creation for the well-being of humans in the form of Vedas.

Knowledge of Vedas can be cognized by the practice of Yoga.